Update single-threaded implementation to new senbuf() function.

+++ sbbs3/postmsg.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet user create/post public message routine */
 
-/* $Id: postmsg.cpp,v 1.82 2010/03/06 00:13:04 rswindell Exp $ */
+/* $Id: postmsg.cpp,v 1.81 2009/10/25 03:12:13 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -185,7 +185,7 @@ bool sbbs_t::postmsg(uint subnum, smbmsg_t *remsg, long wm_mode)
 
 	msg_tmp_fname(useron.xedit, str, sizeof(str));
 	if(!writemsg(str,top,title,wm_mode,subnum,touser,&editor)
-		|| (long)(length=(long)flength(str))<1) {	/* Bugfix Aug-20-2003: Reject negative length */
+		|| (long)(length=flength(str))<1) {	/* Bugfix Aug-20-2003: Reject negative length */
 		bputs(text[Aborted]);
 		return(false); 
 	}
