+++ sbbs3/qwk.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet QWK packet-related functions */
 
-/* $Id: qwk.cpp,v 1.53 2010/03/06 00:13:04 rswindell Exp $ */
+/* $Id: qwk.cpp,v 1.52 2009/12/04 06:04:11 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -694,7 +694,7 @@ void sbbs_t::qwk_sec()
 				continue; 
 			}
 
-			l=(long)flength(str);
+			l=flength(str);
 			bprintf(text[FiFilename],getfname(str));
 			bprintf(text[FiFileSize],ultoac(l,tmp));
 			if(l>0L && cur_cps)
